pyproject.toml
[tool.poetry]
name = "codes"
version = "0.1.0"
description = "Benchmarking tool for coupled ODE surrogate models"
authors = ["Robin Janssen <robin.janssen@stud.uni-heidelberg.de>", "Immanuel Sulzer"]
license = "GPL-3.0"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"
]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
h5py = "^3.9.0"
matplotlib = "^3.8.4"
PyYAML = "^6.0.1"
scipy = "^1.14.1"
tabulate = "^0.9.0"
torch = "^2.3.0"
torchode = "0.2.0"
tqdm = "^4.66.4"
schedulefree = "^1.4"
optuna = "^4.1.0"
sqlalchemy = "^2.0.37"
psycopg2-binary = "^2.9.10"
optuna-dashboard = "^0.17.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pre-commit = "^3.8.0"
pytest = "^8.3.3"
pytest-cov = "^4.0.0"
sphinx = "8.0.2"
sphinx-book-theme = "1.1.3"
sphinxcontrib-napoleon = "^0.7"
sphinx-autodoc-typehints = "2.5.0"


[tool.pytest.ini_options]
testpaths = [
    "test"
]
pythonpath = "."
python_functions = "test_*"
python_classes = "Test*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
